+++ src/glob.cpp
@@ -27,6 +27,7 @@ namespace arsh {
 // ################################
 
 GlobPatternScanner::Status GlobPatternScanner::match(const char *name, const Glob::Option option,
+                                                     const bool allowEmptyPattern,
                                                      std::string *err) {
   // ignore starting with '.'
   if (*name == '.') {
@@ -46,6 +47,10 @@ GlobPatternScanner::Status GlobPatternScanner::match(const char *name, const Glo
         return Status::UNMATCHED_DOT_PREFIX;
       }
     }
+  }
+
+  if (allowEmptyPattern && this->isEndOrSep()) {
+    return Status::MATCHED;
   }
 
   const char *oldName = nullptr;
@@ -380,7 +385,7 @@ Glob::Status Glob::invoke(std::string &&baseDir, const char *iter, std::string *
   this->statCount = 0;
   this->readdirCount = 0;
   std::pair<Status, bool> s;
-  for (; !(s = this->match(baseDir, iter, err)).second; popDir(baseDir))
+  for (; !(s = this->match(baseDir, iter, false, err)).second; popDir(baseDir))
     ;
   if (s.first != Status::MATCH) {
     return s.first;
@@ -454,9 +459,9 @@ std::string Glob::resolveBaseDir(const char *&iter) const {
 }
 
 std::pair<Glob::Status, bool> Glob::match(const std::string &baseDir, const char *&iter,
-                                          std::string *err) {
-  if (hasFlag(this->option, Option::GLOBSTAR) && this->tryConsumeDoubleStars(iter)) {
-    return {this->matchRecursive(baseDir, iter, err), true};
+                                          bool allowEmptyPattern, std::string *err) {
+  if (hasFlag(this->option, Option::GLOBSTAR) && this->consumeDoubleStars(iter)) {
+    return {this->matchDoubleStar(baseDir, iter, err), true};
   }
 
   const auto dir = openDir(baseDir.c_str());
@@ -474,7 +479,7 @@ std::pair<Glob::Status, bool> Glob::match(const std::string &baseDir, const char
     }
 
     GlobPatternScanner scanner(iter, this->end());
-    const auto ret = scanner.match(entry->d_name, this->option, err);
+    const auto ret = scanner.match(entry->d_name, this->option, allowEmptyPattern, err);
     switch (ret) {
     case GlobPatternScanner::Status::UNMATCHED:
     case GlobPatternScanner::Status::UNMATCHED_DOT:
@@ -512,7 +517,7 @@ std::pair<Glob::Status, bool> Glob::match(const std::string &baseDir, const char
           return {Status::MATCH, false};
         }
         auto next = scanner.getIter();
-        auto s = this->match(name, next, err);
+        auto s = this->match(name, next, allowEmptyPattern, err);
         if (!s.second) {
           iter = next;
           rewinddir(dir.get());
@@ -561,9 +566,17 @@ static FileType getFileType(DIR *dir, const struct dirent *entry) {
   return FileType::OTHER;
 }
 
-Glob::Status Glob::matchRecursive(const std::string &baseDir, const size_t targetOffset,
-                                  const char *iter, std::string *err) {
+Glob::Status Glob::matchDoubleStar(const std::string &baseDir, const size_t targetOffset,
+                                   const char *const iter, std::string *err) {
   assert(!baseDir.empty());
+  {
+    auto nextIter = iter;
+    auto s = this->match(baseDir, nextIter, true, err);
+    if (s.first != Status::MATCH) {
+      return s.first;
+    }
+  }
+
   const auto dir = openDir(baseDir.c_str());
   if (!dir) {
     return Status::MATCH;
@@ -586,113 +599,16 @@ Glob::Status Glob::matchRecursive(const std::string &baseDir, const size_t targe
     }
 
     const StringRef name = entry->d_name;
-    if (name == "." /*|| name == ".."*/) {
+    if (name == "." || name == "..") {
       continue;
     }
-    const bool dotdot = name == ".." || name == ".";
-    if (name[0] == '.' && !dotdot && !hasFlag(this->option, Option::DOTGLOB) &&
-        (iter == this->end() || *iter != '.')) {
-      continue;
-    }
-    pathBuf.resize(orgBufSize); // trim
-    pathBuf += name;
-    const auto fileType = getFileType(dir.get(), entry);
-    const auto s = this->matchPath(pathBuf, pathBuf.c_str() + targetOffset,
-                                   fileType == FileType::DIR, dotdot, iter, err);
-    if (s == Status::NOMATCH) {
-      continue; // cur-off (ignore directory)
-    }
-    if (s != Status::MATCH) {
-      return s;
-    }
-
-    if (fileType == FileType::LINK) {
-      continue;
-    }
-    if (fileType == FileType::DIR) {
-      if (const auto s = this->matchRecursive(pathBuf, targetOffset, iter, err);
-          s != Status::MATCH) {
+    if (getFileType(dir.get(), entry) == FileType::DIR) {
+      pathBuf.resize(orgBufSize); // trim
+      pathBuf += name;
+      auto s = this->matchDoubleStar(pathBuf, targetOffset, iter, err);
+      if (s != Status::MATCH) {
         return s;
       }
-    }
-  }
-  return Status::MATCH;
-}
-
-static std::string getFirstElementAndMoveChild(const char *&relative) {
-  std::string value;
-  while (*relative) {
-    const char ch = *relative++;
-    if (ch == '/') {
-      break;
-    }
-    value += ch;
-  }
-  return value;
-}
-
-Glob::Status Glob::matchPath(const std::string &path, const char *relative, const bool isDir,
-                             const bool dotdot, const char *iter, std::string *err) {
-  std::string out;
-  out.reserve(path.size());
-  out.append(path.c_str(), relative - path.c_str());
-  if ((iter == this->end() || *iter == '/') && !dotdot) {
-    if (!out.empty() && out.back() != '/') {
-      out += '/';
-    }
-    out += relative;
-    if (iter != this->end() && *iter == '/') {
-      out += '/';
-      ++iter;
-      if (!isDir) {
-        return Status::MATCH;
-      }
-    }
-    relative = "";
-  }
-
-  const char *const oldIter = iter;
-  const char *oldRelative = relative;
-  while (*relative) {
-    const char *cur = relative;
-    std::string name = getFirstElementAndMoveChild(cur);
-    GlobPatternScanner scanner(iter, this->end());
-    switch (scanner.match(name.c_str(), this->option, err)) {
-    case GlobPatternScanner::Status::MATCHED:
-    case GlobPatternScanner::Status::DOT:
-    case GlobPatternScanner::Status::DOTDOT:
-      break;
-    case GlobPatternScanner::Status::UNMATCHED_DOT_PREFIX:
-    case GlobPatternScanner::Status::UNMATCHED_DOT:
-      return Status::NOMATCH; // not match (cut-offf)
-    case GlobPatternScanner::Status::BAD_PATTERN:
-      return Status::BAD_PATTERN;
-    default:
-      iter = oldIter;
-      getFirstElementAndMoveChild(oldRelative);
-      relative = oldRelative;
-      out.clear(); // re-use re-allocated memory
-      out.append(path.c_str(), relative - path.c_str());
-      continue;
-    }
-    out += name;
-    while (true) {
-      if (scanner.consumeSeps() > 0) {
-        out += '/';
-      }
-      if (scanner.consumeDot()) {
-        out += '.';
-      } else {
-        break;
-      }
-    }
-    iter = scanner.getIter();
-    relative = cur;
-  }
-
-  if (iter == this->end()) {
-    if (!this->addResult(std::move(out))) {
-      return Status::LIMIT;
     }
   }
   return Status::MATCH;