+++ src/glob.h
@@ -142,26 +142,26 @@ private:
 
   const char *end() const { return this->pattern.end(); }
 
-  bool tryConsumeDoubleStars(const char *&iter) const {
+  bool consumeDoubleStars(const char *&iter) const {
     unsigned int count = 0;
     auto old = iter;
-    while (iter != this->end()) {
-      if (*iter == '*' && iter + 1 != this->end() && *(iter + 1) == '*') {
-        iter += 2;
-        count++;
+    while (iter != this->end() && *iter == '*' && iter + 1 != this->end() && *(iter + 1) == '*') {
+      iter += 2;
+      if (iter == this->end()) { // **
         old = iter;
-      } else {
+        count++;
         break;
       }
-      if (iter != this->end() && *iter == '/') {
-        iter++;
+      if (*iter == '/') {
+        count++;
+        do {
+          iter++;
+        } while (iter != this->end() && *iter == '/');
         old = iter;
-        if (iter == this->end()) {
-          old = iter - 1;
+        if (iter == this->end()) { // **///  -> remain '/'
+          old--;
           break;
         }
-      } else {
-        break;
       }
     }
     iter = old;
@@ -175,7 +175,8 @@ private:
    */
   std::string resolveBaseDir(const char *&iter) const;
 
-  std::pair<Status, bool> match(const std::string &baseDir, const char *&iter, std::string *err);
+  std::pair<Status, bool> match(const std::string &baseDir, const char *&iter,
+                                bool allowEmptyPattern, std::string *err);
 
   bool addResult(std::string &&path) {
     if (this->consumer && !this->consumer(std::move(path))) {
@@ -185,28 +186,18 @@ private:
     return true;
   }
 
-  /**
-   * recursively match glob (for globstar)
-   * @param baseDir
-   * @param iter
-   * @param err
-   * @return
-   */
-  Status matchRecursive(const std::string &baseDir, const char *iter, std::string *err) {
+  Status matchDoubleStar(const std::string &baseDir, size_t targetOffset, const char *iter,
+                         std::string *err);
+
+  Status matchDoubleStar(const std::string &baseDir, const char *iter, std::string *err) {
     size_t offset = baseDir.size();
     if (baseDir == ".") {
       offset = 0;
     } else if (!baseDir.empty() && baseDir.back() != '/') {
       offset++;
     }
-    return this->matchRecursive(baseDir, offset, iter, err);
+    return this->matchDoubleStar(baseDir, offset, iter, err);
   }
-
-  Status matchRecursive(const std::string &baseDir, size_t targetOffset, const char *iter,
-                        std::string *err);
-
-  Status matchPath(const std::string &path, const char *relative, bool isDir, bool dotdot,
-                   const char *iter, std::string *err);
 };
 
 template <>
@@ -264,10 +255,15 @@ public:
    * @param name
    * must be null terminated
    * @param option
+   * @param allowEmptyPattern
    * @param err
    * @return
    */
-  Status match(const char *name, Glob::Option option, std::string *err);
+  Status match(const char *name, Glob::Option option, bool allowEmptyPattern, std::string *err);
+
+  Status match(const char *name, Glob::Option option, std::string *err) {
+    return this->match(name, option, false, err);
+  }
 
   enum class CharSetStatus : unsigned char {
     MATCH,