+++ CHANGELOG.md
@@ -25,8 +25,9 @@
         - also support character class such as ``alnum``, ``space``
     - always perform tilde expansion before glob expansion
         - if ``failglob`` is disabled and glob expansion failed, return tilde expanded string
-    - experimental support recursive glob (a.k.a ``globstar``)
-        - enabled vis ``globstar`` runtime option
+    - support recursive glob (a.k.a ``globstar``)
+        - enabled via ``globstar`` runtime option
+        - when ``globstar`` option is enabled, always ignore ``fastglob`` option
 - **Breaking Change**: now not preserve exit status during finally/defer block
     - now the following code is valid
       ```